Clean up case and rename it.
