--- gierer-meinhardt/fenics/precice-adapter-config-inhibitor.json
@@ -1,9 +1,0 @@
-{
-  "participant_name": "Inhibitor",
-  "config_file_name": "../precice-config.xml",
-  "interface": {
-      "coupling_mesh_name": "Inhibitor-Mesh",
-      "write_data_name": "Antagonist",
-      "read_data_name": "Substance"
-    }
-}
