--- gierer-meinhardt/precice-config.xml
@@ -1,61 +1,0 @@
-<?xml version="1.0"?>
-
-<precice-configuration>
-
-  <log>
-    <sink filter="%Severity% > debug and %Rank% = 0" format="---[precice] %ColorizedSeverity% %Message%" enabled="true"/>
-  </log>
-
-  <solver-interface dimensions="2">
-
-    <data:scalar name="Antagonist"/>
-    <data:scalar name="Substance"/>
-
-    <mesh name="Activator-Mesh">
-      <use-data name="Antagonist"/>
-      <use-data name="Substance"/>
-    </mesh>
-
-    <mesh name="Inhibitor-Mesh">
-      <use-data name="Antagonist"/>
-      <use-data name="Substance"/>
-    </mesh>
-
-    <participant name="Activator">
-      <use-mesh name="Activator-Mesh" provide="yes"/>
-      <use-mesh name="Inhibitor-Mesh" from="Inhibitor"/>
-      <write-data name="Substance" mesh="Activator-Mesh"/>
-      <read-data name="Antagonist" mesh="Activator-Mesh"/>
-      <mapping:rbf-thin-plate-splines direction="read" from="Inhibitor-Mesh" to="Activator-Mesh" constraint="consistent"/>
-    </participant>
-
-    <participant name="Inhibitor">
-      <use-mesh name="Inhibitor-Mesh" provide="yes"/>
-      <use-mesh name="Activator-Mesh" from="Activator"/>
-      <write-data name="Antagonist" mesh="Inhibitor-Mesh"/>
-      <read-data name="Substance" mesh="Inhibitor-Mesh"/>
-      <mapping:rbf-thin-plate-splines direction="read" from="Activator-Mesh" to="Inhibitor-Mesh" constraint="consistent"/>
-    </participant>
-
-    <m2n:sockets from="Activator" to="Inhibitor" exchange-directory=".."/>
-
-    <coupling-scheme:serial-implicit>
-      <participants first="Activator" second="Inhibitor"/>
-      <max-time value="1.0"/>
-      <time-window-size value="0.1"/>
-      <max-iterations value="100"/>
-      <exchange data="Substance" mesh="Activator-Mesh" from="Activator" to="Inhibitor" />
-      <exchange data="Antagonist" mesh="Inhibitor-Mesh" from="Inhibitor" to="Activator" initialize="true"/>
-      <relative-convergence-measure data="Substance" mesh="Activator-Mesh" limit="1e-5"/>
-      <relative-convergence-measure data="Antagonist" mesh="Inhibitor-Mesh" limit="1e-5"/>
-      <acceleration:IQN-ILS>
-        <data name="Antagonist" mesh="Inhibitor-Mesh"/>
-        <initial-relaxation value="0.1"/>
-        <max-used-iterations value="10"/>
-        <time-windows-reused value="5"/>
-        <filter type="QR2" limit="1e-3"/>
-      </acceleration:IQN-ILS>
-    </coupling-scheme:serial-implicit>
-
-  </solver-interface>
-</precice-configuration>
