+++ volume-coupled-diffusion/fenics/.gitignore
@@ -1,2 +1,4 @@
 venv
 *.pyc
+*.log
+out
