+++ volume-coupled-diffusion/fenics/clean.sh
